yubico: fix tokentype yubico validation against privacyidea
I've run yubicotoken against the privacyidea /ttype/yubikey
and that failed with an invalid signature. This patch fixes
the error - the answer from the validationserver is
line-delimited with \n, not \r.

Testsuite passes after the change, as well as validation
against privacyidea. I've no registered Id with Yubico,
so I didn't test that.