Put client IP in global request object g
We put the client IP at the beginning of the request
in the request object g. This way it is easier
to mangle with the client IP than if it is taken from
request.remote_addr.

At the beginning of the request we could also overwrite
the client IP with X-Forwarded-For

Working on #356