add rewrite to apache and nginx config to provide /wsapi/2.0/verify
Yubico servers use /wsapi/2.0/verify as the path in the
validation URL. Some tools (e.g. Kolab 2fa) let the
user/admin change the api host, but not the rest of
the URL. Let's redirect the api URL to privacyideas
/ttype/yubikey, but we add the configuration only as a comment,
so it's easy to add when needed, but initially inactive.

Until now only the nginx configuration redirected http requests
to https, but not the apache config. Since this might be too
much, uncomment the redirect, but add that option to the
apache configuration as well so users might enable it easily.