Remove superflous if statement
Either the if statement is superflous or makes no sense.

If loginType == FE, then it is true.
If loginType == BE, then the user is either an admin or a user.
If the user is an admin, the statement is true.

But if the user is a user the statement will only be true,
if privacyIDEABackend is not adminOnly.

I.e. the statement will only be FALSE, if:

BE, user is role user, and privacyIDEABackend == adminOnly.

Would this really be the intended behaviour?