Add triggerchallenge API
The triggerchallenge API allows an administrator to trigger a challenge
of a challenge response token (SMS, email) either via the
username or the token serial number.

If the user has more than one token, challenges for all tokens are
triggered.

This way a challenge can be triggered, without the knowledge of the OTP PIN.

Closes #531