Fix the "check_tls_cert" boolean
The checkbox if the TLS certificate should
be verified is stored in the database. If it is not
checked it could return "0" (string zero).
Which in fact is True.
So we need to use the is_true function from the utils
to avoid checking the certificate if the checkbox is not
checked.

Closes #637