Fix user cache behavior if user and realm are passed
Until now, `_get_resolvers` queried the actual resolver in case a User
object was constructed with only login name and realm, even though a
corresponding cache entry existed. Now, the user cache and the actual
resolver are queried subsequently for each resolver in the realm's
prioritized resolver list. This should save us one resolver call as well
as behave properly in the case of multiple resolvers per realm.

Working on #670