fix transaction_id for multi_challenge requests
In create_challenge we get transactionid=None when adding the first
challenge and generate one. The caller gets the transaction ID back
and hands it back for the next/subsequent tokens for the same
request.

When adding the challenge in the database we always used
transaction_id=None, so multiple U2F tokens get different
transacrion_ids for a single request.

Use transaction_id=transactionid, so we have None for the first
and use the generated transaction ID for the other tokens.