Allow pskc export and import of password type token
The export of a password type token can fail due to the
problem that the otpkey (password) of the pw token is
not returned hexlified but as the ordinary password.

This is why we need to differentiate between hotp,totp and pw token type.

The pw token type is encrypted as is and also imported again in the
same way.

Closes #1109