Fix issues with max key length in mariadb
Using privacyIDEA with mariadb (i.e. on Debian 9.4) resulted in a warning
about a key being to long.
The max key length in mariadb (v10.1) is 767 bytes thus limiting unique
varchar columns to 190 characters (with standard utf8 @ 4bytes encoding).
To avoid this limitation, some steps must be taken:
- mariadb must be started with innodb_large_prefix=on and
  innodb_file_format=Barracuda, which is default in mariadb v10.2.2
- tables must be created with row format DYNAMIC

And to simplify installation on Debian, the database user name will be
taken from the configuration as well.

Fixes #967