Removed usage of `get_all_token_users()`
The function `get_all_token_users()` iterates over all tokens and queries
the corresponding resolver for the user. If there are a lot of tokens in
the DB, this can take quite some time. Since we aren't interested in
number of orphaned tokens here, we query the token DB for assigned tokens.

Update documentation and used `is_true()` from `lib.utils` for checking the
Bool-value from the DB.
Also added leading underscore for internal property-functions.