Allow to do add_policy with a list of strings
We decided to also allow setting a list of strings.
This is interesting for boolean policies. Several boolean
policies might trigger, to allow a certain action.
In this case it makes sense, that the admin does not only see
the first of the boolean policies but all boolean policies.

Working on #874