Fix usage of dictionaries
In Python the order of keys in dictionaries is not guaranteed.
This will be enforced in Python 3.3 - 3.7 so we need to fix the code where
we rely on the ordering of the keys.
This is the case where we calculate the signature of a response or the id
of the LDAP Resolver.
In `user.py` this would have randomly changed the output since it can not
be guaranteed that `username` is read before `user` in the `param`
dictionary.
More Information can be found here:
https://portingguide.readthedocs.io/en/latest/dicts.html#changed-key-order

These issues were found by running the tests with Python 2 and setting the
environment variable `PYTHONHASHSEED=random`