Always use the first loginname-attribute as Username
The LDAP Resolver always uses the first login attribute
as internal username. I.e. the User object is created with
the first login attribute, even if the user logs in with
the secondary (or a third...) attribute.
This way we get a consistent audit log.

The audit-info field contains the information, which login attribute
was acually used for login.

Closes #1082