Fix before_after and the tests
Since the before_after now uses get_user_from_param
to also write realms and resolves the the audit_log,
we get a case effect. The realm in a user object is always
lower case. So the realm1A gets converted to realm1a, since
it is written as userrealm and not anymore as the plain realm entry
as before.