Make triggerchallenge HTTP response consistent
The endpoints /validate/check and /validate/triggerchallenge
are able to trigger a challenge. They where not consisitent.
Now both endpoints return the same JSON, when a challenge is triggered.
If there are several C/R tokens, only ONE transaction_id is created,
which makes it easier for the calling application.

Closes #1353