Remove authentication for testing libs
The `setUp()` of the base test case created an admin token request which
slowed down the startup of the tests. Most of the library tests don't need
an admin authentication token since they don't talk to the API.
Therefore a new base class for API test cases creates the authentication
token in its `setUp()` function.