Fix line numbers in logging decorator
The logging decorator did not print the correct line numbers of the
wrapped functions, instead the line number of the decorator functions
was used.
By inspecting the wrapped function and adding extra data to the log
record, we can provide the correct values in the formatter.

This PR also suppresses the output of the compete policies when checking
for a match (only the policy name is given.

The SecureFormatter class now works with printable characters (Unicode).