Fix logic of auth counters and registration token
The check for the auth counters and max_auth_counters is fixed.
The order or checking the counters and increasing the counters (which
implies deleting the registration token) is changed and fixed.

We now first check if the token is allowed to be used for authentication
("check_all") and then increase the counters (which deletes the
registration token).
This way we need to add an offset to the counter checking.

Fixes #1587