Improve the Firebase configuration
The firebase configuration is a bit tricky.
To avoid uploading the wrong json file or entering wrong data
SMSprovider now can have a config-check-method, which is run
after the SMS provider is created.

This way privacyIDEA checks the sanity of the firebase configuration.
It checks if the correct json file is specified and if
the correct project number is configured.

Closes #1592