Add the possibility to autoassign during migration
An additional passthru_assign policy can be used, to
automatically assign a token, when the passthru to the
RADIUS server was successful.

The split_pin_pass method was moved to the lib.utils.
The split_pin_pass in the models.py was removed, since it
is never used in the productive code.

Closes #1619