Fix U2F enrollment for Chrome 72+
In Chrome 72, the base64url decoding behaviour changed.
This could be fixed, if we use the base64_encode function for u2f
registration and not only for signing.

I did not test it in a good way yet. But I noticed that this works in my
scenario with both, Google Chrome and Mozilla Firefox.

resolves #1636