Fix db migration with postgres
PostgreSQL (and Oracle and SQLite) use sequences for autoincrement columns.
If no sequence is given, one will be created automatically which clashes
with the required one in `models.py`
We have to create the required sequence manually but only if the backend
supports it.
Actually this affects all migration scripts which create columns with an
auto_increment attribute.

Fixes #1676