Fix empty audit entry in case of error
In case an error occurred during handling of a request, the error handler
would finalize the audit log entry and clean the audit data.
But after the error handler, the general `after_request` handler was called
as well, which called `finalize_log()` again (with an empty audit_data)
resulting in an empty line in the audit log.

Also remove the `after_request` handler in `validate.py` and `ttype.py` and
use the handler in `before_after.py` since they all do the same.