Enable python 3.5
In order to avoid extra `decode('utf8')` calls in `json.loads()`, a custom
Response class is added in which the `.json` property does not cache the
json data. Thus we can safely use `Response.json` to get the json data. To
write data back to the response, we need to use
`Response.set_data(json.dumps(data))`. This is immediately available as the
`.json` property.