TIQR token enhancements (#1771)
When we unlock the TIQR token with the App and use the wrong PIN/biometric code. Respond
back how many retries we have left before the token is locked/blocked

Fix hexadecimal challenge length for TiQR

Provide the service_displayname in the tiqrauth qrcode. Else we get
unknown identifier in the screen with please unroll first message

Closes #1771