Fix looping over all tokens when using pre-evens
When a pre-event handler is specified, we need to check if a token is
involved, either by checking for a serial or a user. But if no user was
given, `get_tokens()` returned all tokens, which could slow down the
request significantly if there were a lot of tokens.
By checking for a user before issuing `get_tokens()`, the response time
could be decreased by a magnitude in a setup with 14k Tokens.
Fixes #1686