Make `send_file()` more flexible
The `send_file()` function now takes an extra parameter `content_type`
which defaults to 'text/csv'. This will be the mimetype of the response.

Also the 'Content-disposition' header will be set to 'attachment'.
All places where the content type was not 'application/json' are handled.

Some tests now check for headers and content type.