Add pattern for resource name check
Some characters and complete names are not allowed in some input forms
since they either conflict with internal naming schemes or describe a
different resource.
The different patterns are defined in `static/app.js` and used in the
appropriate controllers to add the pattern attribute to the corresponding
input element.
The input element will not be validated unless the pattern matches.

Fixes #1674