Add admin policy restrictions for token parameter
The restriction of TOTP/HOTP token parameters like hashlib, timestep or OTP
length could be enforced for users via policies.
This PR adds the possibility to restrict these parameters for the admin
scope as well.

The system-wide default configuration for the token types is still usable
but will be overwritten by a corresponding policy.
In the Web-UI these parameters are disabled/hidden if a corresponding
policy is set.

If a default system configuration is set, it won't be added to the form
data anymore and thus won't add confusing token info data.

Working on #1566