Avoid password error in sqlresolver
If a editable SQL resolver is defined without a password mapping,
then the line

   del attributes("password")

will cause a Keyerror. Thus it was not possible to define SQL resolvers
without a password. Deleting the password, only if it exists, now
actually allows us to create users in a SQL resolver without a password
mapping.

Fixes #2030