Add QR codes of stores to enrollment page
The enrollment pages for HOTP, TOTP and PUSH now can
contain QR codes pointing to the corresponding stores.
Depending on the number of the qr codes, the original helptext
is scaled to col-md-12, md-9, md-6 or md-3.

The construct is wrapped in a <p>, so that the shadow of
the tokentype dropdown is not overlapped.

Closes #1919