Honor usernames in webui policies.
The webui policies that are called in get_webui_settings are
fetched after the user has logged in. Thus we already know the
username of the logged in user. Thus we can not only match the
realm but also the username of the logged in user to match webui
settings.

Closes #2038