Fix admin policy for uploading tokens
The admin policy for uploading tokens did not
check for the realms, since it matched the wrong realm.
The Upload API takes the parameter "tokenrealms".
We now evaluate the policies with this realm.

Add test case for successful upload into an allowed realm
and failing to upload into a realm, where the user is not allowerd
to upload into.