Read YAML based logging configuration
The python docs suggest using `dictConfig()` instead of `fileConfig()` for
complex logging configurations.
This PR adds the possibility to read logging configuration from a config
file as well as a YAML file. The YAML based configuration will be passed to
`dictConfig()` and takes precedence over a config-file based configuration.
If both files don't exist (or could not be read), the default logging
configuration will be used.

Working on #2059