Replace stateChangeSuccess by transitions.onBefore and onSuccess
In the new angular-ui-router the event stateChangeSuccess has
been removed. The changing of a state is now handled by
hooks to transitions
onBefore and onSuccess.
In the onBefore hook we now need to save the *current* state
and parameters, so that we can return to this state, when needed.

Fixes #2122

In the onSuccess hook we if all Listeners have reloaed.

Fixes #2125