Fix failing tests
Since the order of keys in a dictionary is not fixed, the tests comparing
the resulting log message might fail.
By using json the order will be fixed and we have a common representation
which can be used for further analysis of the message data.
Changed deprecated `assertEquals()` to `assertEqual()`.