Enable deleting validity time with empty string
Currently it is not possible to delete a validity time start/end through
the webUI. Passing an empty string lead to a server error.
This will be fixed with this PR and passing an empty string deletes the
validity start/end.
Also the UI displays a warning if the end date is before the start date.

Fixes #2263