Fix broken response signature on error
If the request fails with an error, the error handler signs the response.
But the after_request handler will also sign the response and is called
after the error handler. The final signature can thus not be verified since
it includes the signature calculated by the error handler and then
overwrites it.
Only the after_request handler should be decorated to sign the response
since it is always called.

Fixes #2267