Make remote_user aware of Split@Sign option
Move the check for Split@Sign into the `split_user()` function. This way it
applies to all calls for (not) splitting the username into user and realm.

Also the remote user policy wasn't checked correctly. If set to `disable`,
the log-in with a remote user was still possible. This is fixed now.

Fixes #1954