Update challenge response behaviour
An authentication request is now deemed to be a response
to a challenge, simply if it contains a transaction_id.
We do not check for an existing DB-challenge entry anymore,
since this entry could have been cleaned up due to time expiry.

So in ``Tokenckass::is_challenge_response`` we do not check for a DB-challenge
anymore. This way in lib/token.py:check_token_list we get a much
longer list  of ``challenge_response_token_list``.

If a user has several challenge response tokens, this way, an
invalid challenge will now increase failcounters for
every token, not only for the one token, that created the challenge.

Closes #2491