Set the startdate at the earliest possible moment
We set the startdate at the earliest possible moment
in before_app_request. This way we always have the
startdate in the g-object.
The Audit-object is now created with the optional parameter
of the startdate.

In additional we move the finalizing of the audit object
to the teardown hook of flask.

Closes #2254