Add username of failed auth in dashboard
The authentication widget contains the list of the
last 15 failed users.

Note: If a user failed to authenticate more than once he will appear
more than once in the list. We will see, if this helps to identify often
failing users or if we should convert the array to a set (which is not
easily possible in javascript).

Closes #2475