Make webauthn tokens work with triggerchallenge
The webauthn decorators pass information into the
request.all_data, which is then moved to the options-dict
to pass this information to the lower layers.

The triggerchallenge actually was missing the code to
move request params to the options dict.

Now all necessary information is passed to the lower layer.

Fixes #2538