Enable UI login with multiple FIDO tokens
If a user had multiple FIDO tokens enrolled (U2F, WebAuthn), a unique
challenge was created for each token during authentication. The client
on the other hand expected only one challenge to sign even for multiple
tokens. Since the client library took the challenge of the first
challenge-response token to sign, only the first token was able to
answer the challenge correctly. The second token always failed since it
signed the wrong challenge.
By using the transaction ID which is passed to each `create_challenge()`
call we can detect if a challenge for a WebAuthn or U2F token has been
created and reuse the actual challenge data.

Closes #2092