Allow changing POOLING_LOOP_TIMEOUT of ldap3
The ldap3 config setting POOLING_LOOP_TIMEOUT is a system wide
setting, so it does not make sense to set this per LDAP resolver.
Thus we add PI_LDAP_POOLING_LOOP_TIMEOUT to pi.cfg, which is
the first LDAP resolver config in pi.cfg.

This commit does not change the default setting of currently
10 seconds of timeout between rounds. However, since the timeout
is also added if only one round is done, it might make sense
to set this to "0" in pi.cfg for performance reasons.

Closes #2662