Remove usage of pkg_resources
The pkg_resources package is kind of deprecated and the use of
importlib_resources and importlib_metadata is encouraged [1].
This PR also adds the version information to the PI-ASCII-art when
running `pi-manage` and determines the migration path from the installed
distribution files. That way we can stamp the database when running
`pi-manage createdb`.

Fixes #2708

[1]: https://setuptools.readthedocs.io/en/latest/pkg_resources.html