Fix enroll-button staying disabled
In case of an error during enrollment the enroll-button
stays disabled. This happens due to the $scope.enrolling
still being "true" in the webUI. We need to set this
to "false" again. To do so, we add a callback-function
for the error-case. This way we can reset the
enrolling variable in the correct scope.

Closes #2717