Fix the ValidateError in case of poll-only
The push token will raise a ValidateError if
the "res" is false, which would usually indicate,
that there is an error when submitting to firebase.

But since we do not submit to firebase there will be no
error, so we default the "res" to True.

For the polling endpoint the administrator sill needs to
set the push_poll_allow policies.

Closes #2720