Use clientwait with U2F and WebAuthn
We also use the rollout_state "clientwait" for
webauthn and u2f tokens. This way the token janitor
can easily find tokens, that have not been successfully/finally
enrolled.

The logic has changed slightly, so that some testcases would
not run correctly anymore, seemingly the lib-testcases do
not reflect the real world use case in this scenario.

This is why we need to failsafe the name of the
token in the description.

Added a class ROLLOUTSTATE for future use and
to avoid typos in the state names.

Working on #2784