Reduce the code for the SMTP URI
Add the ability to test for SMTP_SSL
in the smtpmock.
When using the class SMTP_SSL this is not
covered by the mock, so we take a detour through
the actual SMTP_SSL class, which in turn calls
our mock SMTP class.
To ensure, that we do not loose parameters, we need
to change the explicit parameters to *args and **kwargs.

Closes #2568