Fix removing of tokenowner when registration token is deleted
The token delete mechanism is cleaned up, so that all information
is deleted on the database model level.
If a token is deleted we take care to delete all relevant information
like tokenowner, tokenrealm...

Add test for checking if the tokenowner entry is removed after the
registration token is used and implicitly deleted.

Closes #2907