Fix NEW routes of most config objects
The route/state of creating an object
was broken. Previouly it was the edit-state
without a given object parameter.
This some either broke or never worked correctly.

I added an additional state like
  config.radius.add
to explicitly reference, when a new configuration
would be created.
This fixed the problem with the link and the
problem with the highlighting.

Closes #2546