Fix disabled otpauth URI scheme
Due to the sanitization of HTML data the otpauth URI-scheme was marked
as disabled.
Adding this together with the netknights.it URL to the trusted URL list
fixes this.
Also in case an administrator changes the GDPR link via policy, the
link-contents must be marked as untrusted to avoid disabling the link.
Fixed some https links as well.

Closes #3067