Change the order of postpolicy and event decorator
At the end of the request the postpolicy has to run before the event
decorator, so that the "check_verify_enrollment" already has set
the rollout_state of the token object. Otherwise the post-event-handler
will not be able to use the rollout_state of the token, since at this
point the rollout state is not set, yet.

Working on #2158