Remove unnecessary hexlification
The webauthn code does an additional hexlification of the credential id
before saving it to the db. Since it is saved in the `key_enc` column
of the token table which is limited to 1024 characters, the maximum
length of the credential id is 512 bytes (while the draft WebAuthn 3
standard proposes a length of 1023 bytes).

In the long run we could save the credential id as a tokeninfo value to
avoid db limits.

Also fixes an unnecessary warning in the `check_verify` postpolicy.

Fixes #3137