Add encoding check to SMPP provider
In order to allow SMS with non-ascii characters, the SMPP library
offers a check for the given encoding.
Also print the `repr()` of an error since just the value of a `KeyError`
is not very meaningful.
And the return values of the SMPP library are added to the debug log.

Closes #3321