Make realm in auth-request case insensitive
While the user is found in the correct resolver/realm (by changing the
realm to lower case), when querying for the admin rights the original
realm was used and so no policy matched the realm and the admin didn't
have any rights.
Fixing this by changing the realm to lower case.

Closes #3329