Fix action parameter when saving policy
In some cases when a policy with an `int` type is saved in the UI the
wrong action value was send to the API. This was due to the handling of
the multi-value actions introduced in #3280.
Since we have the types of each value in the `$scope.actions` array we
can use this to create the correct parameter for the action value.

Closes #3347