Fix broken login from UI
When a user logs in via the WebUI, there is always a `realm` parameter
added to the `/auth` request, even if no realm is given (empty string).
With the changes from fa643349 this lead to the user not being found if
the login was `user@realm` and `realm` wasn't the default realm.

Closes #3416