Check token assignment before deleting tokengroups
Deleting a tokengroup which has still tokens assigned to it, now fails
with an error.
Also when deleting a token, the tokengroup assignment will be deleted as
well.
And a bug in the UI has been fixed which prevented the "save tokengroup"
button to be displayed in the token-view.

Fixes #3423