Ignore PIN policies during rollover and verify
The "rollover" and "verify" actions both call the `/token/init` endpoint
which also checks if the given token PIN matches the PIN policies of the
logged in user.
However during "rollover" or "verify" there is no PIN setting necessary
so we exclude these `/token/init` variants from the PIN checking.

This PR also removes the PIN entry field in case a user initiates a
token rollover.

Closes #2886