Fix calculation of TOTP values
Instead of using the `floor` function for calculating the counter of the
*TOTP* algorithm, the counter was rounded to the nearest integer value.
Thus the validity of a single *OTP* value was always shifted `timestep /
2` into the future.
For the regular *TOTP* algorithm this wasn't a problem since it also
defines a time-drift and a time-window for checking the value.
The *DayPassword* token which is derived from the *TOTP* algorithm
however does not use a time-window.

This fix should not severely affect existing *TOTP* tokens since the
algorithm already accounts for time-drifts.

Closes #3734