Update migration script order
To avoid merges in the database migration scripts, the order of the
scripts is adjusted.

To avoid an error about missing tables, the migration script for
revision 5cb310101a1f has also been adjusted to check if tables exist.
This migration script is part of the 3.9 release whose branch hasn't
been updated (and probably doesn't need to be).