Fix broken Galera-Cluster installation
Galera cluster based on MariaDB >= 10.6 was unable to create the initial
tables since it needed sequences with an increment of 0.
Unfortunately PostgreSQL does not work with an increment of 0 in
sequences.
Compiling the SQL statements for creating a sequence differently for
MySQL/MariaDB and PostgreSQL fixes this.
Database upgrades to this version still work as expected since the
increment was already used in the migration script (now removed).

Fixes #3863