Fixes the output of the timestep of the TOTP export
When running

  privacyidea-token-janitor find --type totp --action export --csv

the last column of the TOTP token is not filled.
This is due to there is no "timestep" in token_dict.
So we pull the "timeStep" from the
token_dict.get("info_list").

This way we have all token information to use the csv export
for later import.