Handle undefined action values in JS
In case of an unknown action value which isn't in the `allowedValues`
list, the processing of the action failed and the policy action list
isn't rendered properly.
Ignoring an unknown action value fixes this.
This was the case i.e. when `email` was added to the
`challenge_response` action value in a version < v3.9 but the
email-token is a challenge-response token by default and thus was
removed from the list of token types where challenge-response was an
alternative.