Add version information to export data
When exporting data via `pi-manage config export` and additional entry
with the current privacyIDEA version is added.
During import through `pi-manage config import` it is checked whether
the version information is present and if it matches the current
version. If not, a warning is printed to check the resulting
configuration in privacyIDEA.
Also added some error handling during import and introduces some pytest
style testing with fixtures.

Closes #4055