Remove space before regtoken on double-click copy
When the registration token is double-clicked and copied, a space is
added to the front of string in the clipboard, leading to " TOKEN"
instead of "TOKEN". When copied blindly (e.g. during testing),
this causes an invalid token and possibly confusion.
This does not happen on Chromium (tested on Firefox 130.0.1 and
Chromium 129.0.6668.58).