PKCS12 export on pending request
If the certificate signing request is pending (MSCA), only the private
key is exported in the PKCS12 container. The certificate has to be
included manually when the signing request is finished.

The enrolled page for the certificate token prominently warns about the
availability of the generated PKCS12 password.

Also improved documentation and removed some unnecessary code.