Fix database rollback between tests
Some tests require that data from a previous test is available in the
database so we need to commit all changes in the `tearDown()` function
of the tests. In case of an error the changes are rolled back to allow
following tests to succeed.

And split up the `test_api_validate.py` file because it grew to big...