Avoid KeyError when searching for user in SQLResolver
When the column-to-keyword mapping in the SQLIdResolver is wrong, the
search for a user results in a 500 error (The authentication flow is not
affected by this!).
We check the given keywords if the exist in the mapping and if the
requested columns do exist in the user table in the db.
Currently this only results in an error message in the privacyIDEA log
since there could be multiple resolvers to check.

Closes #4461