Check resolver search parameter for existence
If a search parameter for the `getUserList()` functions was not
available in the resolvers mapping, a KeyError was thrown which resulted
in a 500 response from the server.
By checking if the search parameter exists we can issue a proper warning
and let the caller decide whether to continue the search.

Closes: #4541